+++ tests/_data/StubProject/Controller/Custom.php
@@ -7,6 +7,11 @@ use Phalcon\Mvc\Controller;
 
 class Custom extends Controller
 {
+    public function indexAction()
+    {
+        return new Response(__CLASS__ . '::' . __FUNCTION__ . '()');
+    }
+    
     public function fooAction()
     {
         echo __CLASS__ . '::' . __FUNCTION__ . '()';